+++ src/Lynx.Benchmark/Lynx.Benchmark.csproj
@@ -11,7 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Lynx.ConstantGenerator\Lynx.ConstantGenerator.csproj" />
+    <ProjectReference Include="..\Lynx.ConstantsGenerator\Lynx.ConstantsGenerator.csproj" />
     <ProjectReference Include="..\Lynx\Lynx.csproj" />
   </ItemGroup>
