+++ src/Lynx.ConstantsGenerator/BitBoardsGenerator.cs
@@ -1,11 +1,14 @@
 ﻿using Lynx.Model;
 
-namespace Lynx.ConstantGenerator;
+namespace Lynx.ConstantsGenerator;
 
 #pragma warning disable S3353 // Unchanged local variables should be "const" - FP https://community.sonarsource.com/t/fp-s3353-value-modified-in-ref-extension-method/132389
 #pragma warning disable S106, S2228 // Standard outputs should not be used directly to log anything
 
-public static class ConstantsGenerator
+/// <summary>
+/// See layout 3 in https://tearth.dev/bitboard-viewer/
+/// </summary>
+public static class BitBoardsGenerator
 {
     public static BitBoard NotAFile()
     {
@@ -109,38 +112,6 @@ public static class ConstantsGenerator
         for (int rank = 8; rank >= 1; --rank)
         {
             Console.WriteLine($"a{rank}, b{rank}, c{rank}, d{rank}, e{rank}, f{rank}, g{rank}, h{rank},");
-        }
-    }
-
-    public static void BishopRelevantOccupancyBits()
-    {
-        for (var rank = 0; rank < 8; ++rank)
-        {
-            for (var file = 0; file < 8; ++file)
-            {
-                int square = BitBoardExtensions.SquareIndex(rank, file);
-
-                var bishopOccupancy = AttackGenerator.MaskBishopOccupancy(square);
-                Console.Write($"{bishopOccupancy.CountBits()}, ");
-            }
-
-            Console.WriteLine();
-        }
-    }
-
-    public static void RookRelevantOccupancyBits()
-    {
-        for (var rank = 0; rank < 8; ++rank)
-        {
-            for (var file = 0; file < 8; ++file)
-            {
-                int square = BitBoardExtensions.SquareIndex(rank, file);
-
-                var bishopOccupancy = AttackGenerator.MaskRookOccupancy(square);
-                Console.Write($"{bishopOccupancy.CountBits()}, ");
-            }
-
-            Console.WriteLine();
         }
     }
 }
