+++ src/Lynx.Dev/Lynx.Dev.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Lynx.ConstantGenerator\Lynx.ConstantGenerator.csproj" />
+    <ProjectReference Include="..\Lynx.ConstantsGenerator\Lynx.ConstantsGenerator.csproj" />
     <ProjectReference Include="..\Lynx\Lynx.csproj" />
   </ItemGroup>
