+++ src/Lynx.Dev/Program.cs
@@ -1,5 +1,5 @@
 ﻿using Lynx;
-using Lynx.ConstantGenerator;
+using Lynx.ConstantsGenerator;
 using Lynx.Model;
 using System.Diagnostics;
 using System.Numerics;
