+++ src/Lynx/Lynx.csproj
@@ -48,7 +48,7 @@
       <_Parameter1>$(AssemblyName).Benchmark</_Parameter1>
     </AssemblyAttribute>
     <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-      <_Parameter1>$(AssemblyName).ConstantGenerator</_Parameter1>
+      <_Parameter1>$(AssemblyName).ConstantsGenerator</_Parameter1>
     </AssemblyAttribute>
     <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
       <_Parameter1>$(AssemblyName).Test</_Parameter1>
